requirements.txt
pystray
flask
opencv-python
waitress
pillow
imutils